feat(ae-239): migrate loading component to typescript
