+++ src/components/Loading/Loading.spec.tsx
@@ -3,10 +3,6 @@ import { render } from '@testing-library/react';
 import { Loading } from './Loading';
 
 describe('<Loading />', () => {
-  it('should be defined', () => {
-    expect(Loading).toBeDefined();
-  });
-
   it('should render base component correctly', () => {
     const { container } = render(<Loading isLoading={true} />);
